Move Segment and BlkDevSegment to shared module
